requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.80.0
h11==0.13.0
httptools==0.5.0
idna==3.3
pydantic==1.10.7
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
watchfiles==0.16.1
websockets==10.3
requests==2.28.1
bs4==0.0.1
selenium==3.141.0
webdriver_manager==3.8.4
urllib3==1.26.8
Pillow
packaging==21.3
lxml==4.9.2
numpy
opencv-python
requires.io